requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.45
Werkzeug==2.2.2
psycopg2-binary==2.9.5
Flask-Migrate==4.0.1
flask-admin==1.6.0
Flask-COMBO-JSONAPI==1.1.0
markupsafe==2.0.1
ComboJSONAPI==1.0.5
apispec==2.0.2
PyYAML==5.3.1
python-dotenv==0.17.1